break apart teams.hpp into backends
